Refactoring texture-readback and making sure it works on all platforms with different alignment requirements
Former-commit-id: 825c8f033f81935c20f59885fedd801980bbe4c7 [formerly 54e8888ff81ff59744e924f19754dbf50c0b679a]
Former-commit-id: 7d036166864aef8eb320ebe078d11b257e6665cc

+++ fancy_core/CommandList.cpp
@@ -167,7 +167,7 @@ namespace Fancy {
     }
 
     if (ringBufferList->empty() || ringBufferList->back()->GetFreeDataSize() < aDataSize)
-      ringBufferList->push_back(RenderCore::AllocateRingBuffer(CpuMemoryAccessType::CPU_WRITE, bindFlags, MathUtil::Align(aDataSize, sizeStep), name.c_str()));
+      ringBufferList->push_back(RenderCore::AllocateRingBuffer(CpuMemoryAccessType::CPU_WRITE, bindFlags, (uint) MathUtil::Align(aDataSize, sizeStep), name.c_str()));
 
     GpuRingBuffer* ringBuffer = ringBufferList->back();
     uint64 offset = 0;
